requirements.txt
openai>=1.30.0
Pillow>=10.3.0
pyyaml>=6.0
rich>=13.7.0
click>=8.1.0
pydantic>=2.7.0
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
